fix bug in bh_sampling
some of the minima in the database don't satisfly the configuration
check (they are extended structures with some of the atoms outside a
sphere of radius 2.5).  I now check for this properly now (this check
was called improperly), and resample until I get a configuration that
passes the configuration checks.