simplify NestedSampling by making mc_runner mandatory
then you don't need to pass a takestep function or accept tests

stepsize is now an attribute of nested sampling, not of takestep

fix a bug in bh_sampling where minprob was not being passed correctly.
I think mciter was being passed as minprob.  This might have meant mciter
was taking the default value.